fix string tensor for non ascii char.

+++ test/TensorFlowNET.UnitTest/ManagedAPI/LinalgTest.cs
@@ -13,9 +13,9 @@ namespace TensorFlowNET.UnitTest.ManagedAPI
 
             Assert.AreEqual((3, 3), tensor.TensorShape);
 
-            Assert.AreEqual(0.0f, (float)tensor[2, 0]);
-            Assert.AreEqual(0.0f, (float)tensor[2, 1]);
-            Assert.AreEqual(1.0f, (float)tensor[2, 2]);
+            Assert.AreEqual(0.0f, (double)tensor[2, 0]);
+            Assert.AreEqual(0.0f, (double)tensor[2, 1]);
+            Assert.AreEqual(1.0f, (double)tensor[2, 2]);
         }
     }
 }
